Move code to drizzle.py
